+++ CHANGELOG.md
@@ -63,3 +63,9 @@
 ### Adicionando
 
 - Conexão de botão de editar avaliação com modal de edição
+
+## US6.4 - PR 11
+
+### Adicionando
+
+- Modal de detalhar avaliações